+++ bot/mod_commands/listservers.js
@@ -9,8 +9,11 @@ module.exports = {
 		let sendMsg = [];
 		let arr = bot.guilds.array()
 		sendMsg.push(`Currently watching over ${arr.length} servers:\n`);
-		for (let i = 0; i < arr.length; i++) {
-			sendMsg.push(`[${i}] [${arr[i].id}] ${arr[i].name}`);
+		//Only spit out detailed list while in DM/PM channel
+		if (!msg.guild) {
+			for (let i = 0; i < arr.length; i++) {
+				sendMsg.push(`[${i}] [${arr[i].id}] ${arr[i].name}`);
+			}
 		}
 		utils.sendLongMessage(bot, msg, sendMsg.join("\n"), true, "\n");
 	}